Refactor: Migrate to AGP 9.0 project structure
Migrate composeApp and data modules from the legacy android plugins to `com.android.kotlin.multiplatform.library`, and extract the Android application entry-point into a dedicated androidApp module.
