+++ .github/workflows/build.yml
@@ -17,7 +17,7 @@ jobs:
       - name: Setup Gradle
         uses: gradle/actions/setup-gradle@v4.0.0
       - name: Build Android App
-        run: ./gradlew composeApp:assembleDebug
+        run: ./gradlew androidApp:assembleDebug
       - name: Run JVM Tests
         run: ./gradlew domain:jvmTest
   ios:
