+++ androidApp/src/debug/res/values/strings.xml
