+++ androidApp/src/development/res/values/strings.xml
