+++ androidApp/src/main/res/drawable/abysner_logo.xml
