+++ gradle.properties
@@ -2,6 +2,10 @@ kotlin.code.style=official
 
 #Gradle
 org.gradle.jvmargs=-Xmx4096M -Dfile.encoding=UTF-8 -Dkotlin.daemon.jvm.options\="-Xmx4096M"
+org.gradle.java.installations.auto-detect=true
+org.gradle.java.installations.auto-download=true
+org.gradle.daemon.toolchain.version=21
+org.gradle.daemon.toolchain.vendor=ADOPTIUM
 
 #Android
 android.nonTransitiveRClass=true
@@ -9,3 +13,10 @@ android.useAndroidX=true
 
 #Kotlin Multiplatform
 kotlin.mpp.enableCInteropCommonization=true
+
+#Abysner
+# DMG distribution does not support "-beta", MSI requires at least MAJOR.MINOR.BUILD
+abysnerVersionBase=1.0.8
+abysnerVersion=1.0.8-beta
+# iOS supports a String here, but Android only an integer
+abysnerBuildNumber=10
