+++ iosApp/Configuration/Config.xcconfig
@@ -1,3 +1,4 @@
+#include "Version.xcconfig"
 TEAM_ID=
 BUNDLE_ID=nl.neotech.app.abysner
 APP_NAME=Abysner
