+++ iosApp/Configuration/Version.xcconfig
@@ -1,0 +1,2 @@
+MARKETING_VERSION=1.0.8-beta
+CURRENT_PROJECT_VERSION=10
